requirement.txt
Flask
gunicorn
pandas
sklearn
numpy
sklearn
scikit-learn
PyYAML
matplotlib
six
-e .